(hopefully) made it possible to run through double click
Now makes the user select the output file path because it wasn't possible to run through double click, as it made use of a relative file path. Got rid of the original output.txt as its no longer useful. Also fixed a mistake where the title for the window didn't show up due to a mistake in syntax.